Moved test files next to their counterpart
