+++ src/utils/formatString.test.ts
